can rerun existing hugo site

+++ hugo/hugostart.sh
@@ -1,31 +1,44 @@
 #!/bin/sh
 set -e
 
-#Creating clean hugo image in ../hugodir
-echo "Moving to hugodir"
-cd ../hugodir
-echo "Running Hugo new site "
-hugo new site . --force
+
+if [ ! -d "../hugodir/content" ] 
+then
+    #Creating clean hugo image in ../hugodir
+    echo "Moving to hugodir"
+    cd ../hugodir
+    echo "Running Hugo new site "
+    hugo new site . --force
 
 
-#Remove default folders
-rm -R content/
-rm -R config.toml
+    #Remove default folders
+    rm -R content/
+    rm -R config.toml
 
-#Clonning Bento Theme
-cd themes/
-git clone https://github.com/leonardofaria/bento.git
+    #Clonning Bento Theme
+    cd themes/
+    git clone https://github.com/leonardofaria/bento.git
 
-#Moving directories to hugodir/
-cp ./bento/package.json ../
-mv ./bento/exampleSite/content ../
-mv ./bento/exampleSite/config.toml ../
+    #Moving directories to hugodir/
+    cp ./bento/package.json ../
+    mv ./bento/exampleSite/content ../
+    mv ./bento/exampleSite/config.toml ../
 
-#Installing package.json
-cd ..
-npm i
+    #Installing package.json
+    cd ..
+    npm i
 
-#Starting hugo server
-hugo
-hugo server
+    #Starting hugo server
+    hugo
+    hugo server
+else 
+    echo "Moving to hugodir"
+    cd ../hugodir
+
+    #Starting hugo server
+    hugo
+    hugo server
+fi
+
+
